# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MMSv2",
        "description": "Monospace Member System API 文件",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Paginate all users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Number of users per page",
                        "required": false,
                        "example": "10"
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter users by name or email or mobile",
                        "required": false,
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "custom_id": {
                                        "type": "string",
                                        "default": "7E0001"
                                    },
                                    "username": {
                                        "type": "string",
                                        "default": "test01"
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "default": "test01"
                                    },
                                    "email": {
                                        "type": "string",
                                        "default": "test01@example.com"
                                    },
                                    "mobile_country_code": {
                                        "type": "string",
                                        "default": "TW"
                                    },
                                    "mobile_country_calling_code": {
                                        "type": "string",
                                        "default": "+886"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "default": "905123456"
                                    },
                                    "password": {
                                        "type": "password",
                                        "default": "password"
                                    },
                                    "confirm_password": {
                                        "type": "password",
                                        "default": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Display User by user id",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ffffffff-ffff-ffff-ffff-ffffffffffff"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update User info by user id",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ffffffff-ffff-ffff-ffff-ffffffffffff"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "default": "test"
                                    },
                                    "email": {
                                        "type": "string",
                                        "default": "test@example.com"
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "default": "Test"
                                    },
                                    "mobile_country_code": {
                                        "type": "string",
                                        "default": "TW"
                                    },
                                    "mobile_country_calling_code": {
                                        "type": "string",
                                        "default": "+886"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "default": "905123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete User by user id",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ffffffff-ffff-ffff-ffff-ffffffffffff"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/users/{id}/organization": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create/Update the User Organization by user id",
                "operationId": "2d458465a7f2faf095cc87eaab88ceaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ffffffff-ffff-ffff-ffff-ffffffffffff"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": "測試股份有限公司"
                                    },
                                    "vat": {
                                        "type": "string",
                                        "default": "0000000"
                                    },
                                    "phone_country_code": {
                                        "type": "string",
                                        "default": "TW"
                                    },
                                    "phone_country_calling_code": {
                                        "type": "string",
                                        "default": "+886"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "default": "0422000000"
                                    },
                                    "email": {
                                        "type": "string",
                                        "default": "test@example.com"
                                    },
                                    "address": {
                                        "properties": {
                                            "city": {
                                                "type": "string",
                                                "default": "台中市"
                                            },
                                            "region": {
                                                "type": "string",
                                                "default": "西區"
                                            },
                                            "zip_code": {
                                                "type": "string",
                                                "default": "403"
                                            },
                                            "address_line_1": {
                                                "type": "string",
                                                "default": "台灣大道一段"
                                            },
                                            "address_line_2": {
                                                "type": "string",
                                                "default": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "406": {
                        "description": "Only Accept application/json"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}